school_in_a_box/requirements.txt
fastapi 
uvicorn 
langgraph 
pymongo 
python-dotenv
pytesseract